+++ crates/mimic/src/interface/query.rs
@@ -2,10 +2,9 @@ use crate::{
     Error,
     core::Key,
     db::query::{DeleteQuery, LoadQuery, SaveQuery},
-    interface::InterfaceError,
 };
 use candid::Principal;
-use icu::cdk::call::Call;
+use icu::{IcuError, cdk::call::Call};
 use thiserror::Error as ThisError;
 
 ///
@@ -14,15 +13,11 @@ use thiserror::Error as ThisError;
 
 #[derive(Debug, ThisError)]
 pub enum QueryError {
-    #[error("call error: {0}")]
-    CallError(String),
-
     #[error("entity not found: {0}")]
     EntityNotFound(String),
 }
 
 // query_load
-// currently disabled because LoadResponse needs a rethink
 pub async fn query_load(pid: Principal, query: LoadQuery) -> Result<Vec<Key>, Error> {
     query_call(pid, "mimic_query_load", &query).await
 }
@@ -47,13 +42,9 @@ async fn query_call<T: candid::CandidType + for<'de> candid::Deserialize<'de>>(
     let result = Call::unbounded_wait(pid, method)
         .with_arg(arg)
         .await
-        .map_err(|e| QueryError::CallError(e.to_string()))
-        .map_err(InterfaceError::QueryError)?;
+        .map_err(IcuError::from)?;
 
-    let response = result
-        .candid::<T>()
-        .map_err(|e| QueryError::CallError(e.to_string()))
-        .map_err(InterfaceError::QueryError)?;
+    let response = result.candid::<T>().map_err(IcuError::from)?;
 
     Ok(response)
 }