+++ crates/mimic_build/src/actor/mod.rs
@@ -1,6 +1,7 @@
 pub mod db;
 pub mod fixtures;
 pub mod query;
+pub mod stats;
 
 use mimic_schema::{
     build::get_schema,
@@ -49,6 +50,7 @@ impl ActorBuilder {
 
         // shared between all canisters
         tokens.extend(db::generate(&self));
+        tokens.extend(stats::generate(&self));
         tokens.extend(fixtures::generate(&self));
         tokens.extend(query::generate(&self));